+++ modules/mauth-authenticator-apachehttp/src/main/java/com/mdsol/mauth/apache/HttpClientPublicKeyProvider.java
@@ -2,12 +2,14 @@ package com.mdsol.mauth.apache;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.github.benmanes.caffeine.cache.Caffeine;
+import com.github.benmanes.caffeine.cache.Expiry;
 import com.github.benmanes.caffeine.cache.LoadingCache;
 import com.mdsol.mauth.AuthenticatorConfiguration;
 import com.mdsol.mauth.Signer;
 import com.mdsol.mauth.exception.HttpClientPublicKeyProviderException;
 import com.mdsol.mauth.util.MAuthKeysHelper;
 import com.mdsol.mauth.utils.ClientPublicKeyProvider;
+
 import org.apache.http.*;
 import org.apache.http.client.ResponseHandler;
 import org.apache.http.client.methods.HttpGet;
@@ -35,48 +37,43 @@ public class HttpClientPublicKeyProvider implements ClientPublicKeyProvider {
   private final Signer signer;
   private final CloseableHttpClient httpclient;
   private final PublicKeyResponseHandler publicKeyResponseHandler;
-
-  private LoadingCache<UUID, PublicKey> publicKeyCache;
-  private Long ttl;
+  private final LoadingCache<UUID, PublicKeyData> publicKeyCache;
 
   public HttpClientPublicKeyProvider(AuthenticatorConfiguration configuration, Signer signer) {
     this.configuration = configuration;
     this.signer = signer;
     this.httpclient = HttpClients.createDefault();
     this.publicKeyResponseHandler = new PublicKeyResponseHandler();
+    this.publicKeyCache = setupCache();
   }
 
-  private void setupCache() {
-    publicKeyCache =
-      Caffeine.newBuilder()
-        .expireAfterAccess(ttl, TimeUnit.SECONDS)
+  private LoadingCache<UUID, PublicKeyData> setupCache() {
+    return Caffeine.newBuilder()
+        .expireAfter(new Expiry<UUID, PublicKeyData>() {
+          public long expireAfterCreate(UUID key, PublicKeyData data, long currentTime) {
+            return TimeUnit.SECONDS.toNanos(data.getMaxAgeSeconds());
+          }
+          public long expireAfterUpdate(UUID key, PublicKeyData data, long currentTime, long currentDuration) {
+            return currentDuration;
+          }
+          public long expireAfterRead(UUID key, PublicKeyData data, long currentTime, long currentDuration) {
+            return currentDuration;
+          }
+        })
         .build(this::getPublicKeyFromMauth);
   }
 
-  private PublicKey getPublicKeyFromMauth(UUID appUUID) {
+  private PublicKeyData getPublicKeyFromMauth(UUID appUUID) {
     byte[] payload = new byte[0];
     String requestUrlPath = getRequestUrlPath(appUUID);
     Map<String, String> headers = signer.generateRequestHeaders("GET", requestUrlPath, payload, "");
     String requestUrl = configuration.getBaseUrl() + requestUrlPath;
-    String publicKeyAsString = get(requestUrl, headers, publicKeyResponseHandler);
-    return MAuthKeysHelper.getPublicKeyFromString(publicKeyAsString);
+    return get(requestUrl, headers, publicKeyResponseHandler);
   }
 
   @Override
   public PublicKey getPublicKey(UUID appUUID) {
-    try {
-      if (publicKeyCache == null) {
-        // Lazy load public key cache so that we can set the ttl based on the first response max-age
-        // Do Eureka call first to set the ttl
-        PublicKey key = getPublicKeyFromMauth(appUUID);
-        setupCache();
-        publicKeyCache.put(appUUID, key);
-      }
-      return publicKeyCache.get(appUUID);
-    } catch (Exception e) {
-      logger.error("Public key retrieval error", e);
-      throw new HttpClientPublicKeyProviderException(e);
-    }
+    return publicKeyCache.get(appUUID).getPublicKey();
   }
 
   private String getRequestUrlPath(UUID appUUID) {
@@ -91,25 +88,26 @@ public class HttpClientPublicKeyProvider implements ClientPublicKeyProvider {
       }
       return httpclient.execute(httpGet, responseHandler);
     } catch (IOException ex) {
+      logger.error("Public key retrieval error", ex);
       throw new HttpClientPublicKeyProviderException(ex);
     }
   }
 
-  private class PublicKeyResponseHandler implements ResponseHandler<String> {
+  private class PublicKeyResponseHandler implements ResponseHandler<PublicKeyData> {
     private static final String MAX_AGE = "max-age";
     private static final String PUBLIC_KEY_STR = "public_key_str";
 
     @Override
-    public String handleResponse(HttpResponse response) throws IOException {
+    public PublicKeyData handleResponse(HttpResponse response) throws IOException {
       if (response.getStatusLine().getStatusCode() == HttpStatus.SC_OK) {
-        if (ttl == null) {
-          ttl = getMaxAge(response).orElse(configuration.getTimeToLive());
-        }
+        Long timeToLive = getMaxAge(response).orElse(configuration.getTimeToLive());
 
         HttpEntity entity = response.getEntity();
         String responseAsString = EntityUtils.toString(entity, StandardCharsets.UTF_8);
         ObjectMapper mapper = new ObjectMapper();
-        return mapper.readTree(responseAsString).findValue(PUBLIC_KEY_STR).asText();
+        String publicKeyString = mapper.readTree(responseAsString).findValue(PUBLIC_KEY_STR).asText();
+
+        return new PublicKeyData(MAuthKeysHelper.getPublicKeyFromString(publicKeyString), timeToLive);
       } else {
         throw new HttpClientPublicKeyProviderException("Invalid response code returned by server: "
           + response.getStatusLine().getStatusCode());
@@ -125,4 +123,23 @@ public class HttpClientPublicKeyProvider implements ClientPublicKeyProvider {
           .findFirst());
     }
   }
+
+  private static class PublicKeyData {
+    private final PublicKey publicKey;
+    private final Long maxAgeSeconds;
+
+    public PublicKeyData(PublicKey publicKey, Long maxAgeSeconds) {
+      this.publicKey = publicKey;
+      this.maxAgeSeconds = maxAgeSeconds;
+    }
+
+    public PublicKey getPublicKey() {
+      return publicKey;
+    }
+
+    public Long getMaxAgeSeconds() {
+      return maxAgeSeconds;
+    }
+  }
+
 }